+++ app/scripts/IdHorizontal1DTiledPixiTrack.js
@@ -1,5 +1,7 @@
-import * as PIXI from 'pixi.js';
 import HorizontalTiled1DPixiTrack from './HorizontalTiled1DPixiTrack';
+
+// Configs
+import { GLOBALS } from './configs';
 
 class IdHorizontal1DTiledPixiTrack extends HorizontalTiled1DPixiTrack {
   constructor(context, options) {
@@ -21,11 +23,9 @@ class IdHorizontal1DTiledPixiTrack extends HorizontalTiled1DPixiTrack {
      */
 
     const graphics = tile.graphics;
-    tile.textGraphics = new PIXI.Graphics();
-    // tile.text = new PIXI.Text(tile.tileData.zoomLevel + "/" + tile.tileData.tilePos.join('/')
-    // + '/' + tile.mirrored,
+    tile.textGraphics = new GLOBALS.PIXI.Graphics();
 
-    tile.text = new PIXI.Text(
+    tile.text = new GLOBALS.PIXI.Text(
       `${tile.tileData.zoomLevel}/${tile.tileData.tilePos.join('/')}`,
       {
         fontFamily: 'Arial',