+++ test/GeneAnnotationsTrackTests.js
@@ -49,28 +49,32 @@ describe('Gene Annotations Tracks', () => {
   it('clicks on a gene', done => {
     let clicked = 0;
 
-    api.on('click', data => {
-      expect(clicked).to.eql(0);
-      clicked += 1;
-    });
+    try {
+      api.on('click', () => {
+        expect(clicked).to.eql(0);
+        clicked += 1;
+      });
 
-    waitForTilesLoaded(api.getComponent(), () => {
-      const canvasElem = div.querySelector('canvas');
-      const loc = {
-        clientX: 237,
-        clientY: 117,
-        screenX: 278,
-        screenY: 231
-      };
+      waitForTilesLoaded(api.getComponent(), () => {
+        const canvasElem = div.querySelector('canvas');
+        const loc = {
+          clientX: 237,
+          clientY: 117,
+          screenX: 278,
+          screenY: 231
+        };
 
-      canvasElem.dispatchEvent(createPointerEvent('pointerdown', loc));
-      canvasElem.dispatchEvent(createPointerEvent('pointerup', loc));
+        canvasElem.dispatchEvent(createPointerEvent('pointerdown', loc));
+        canvasElem.dispatchEvent(createPointerEvent('pointerup', loc));
 
-      // console.log('clicked:', clicked);
-      expect(clicked).to.eql(1);
+        // console.log('clicked:', clicked);
+        expect(clicked).to.eql(1);
 
-      done();
-    });
+        done();
+      });
+    } catch (e) {
+      console.warn(e);
+    }
   });
 
   it('changes the color of the minus strand', done => {